requirements.txt
altair==5.1.1
annotated-types==0.5.0
anyio==3.3.4
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
Automat==22.10.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bertopic==0.15.0
bleach==4.1.0
blinker==1.6.2
blis==0.7.10
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
colorama==0.4.6
confection==0.1.3
constantly==15.1.0
contourpy==1.1.0
cryptography==38.0.3
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
Cython==0.29.36
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docutils==0.20.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
filelock==3.8.0
fonttools==4.42.1
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
hdbscan==0.8.33
huggingface-hub==0.17.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
itemadapter==0.7.0
itemloaders==1.0.6
jaraco.classes==3.3.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.6
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.4
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==3.0.4
keyring==24.2.0
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.40.1
lxml==4.7.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplot==0.1.9
matplotlib==3.7.3
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
more-itertools==10.1.0
mpmath==1.3.0
murmurhash==1.0.10
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.1
nh3==0.2.14
nltk==3.8.1
notebook==6.4.5
numba==0.57.1
numpy==1.24.4
opencv-python==4.5.4.60
openpyxl==3.0.9
outcome==1.2.0
packaging==21.2
pandas==1.3.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.2
pathlib==1.0.1
pathy==0.10.2
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.9.6
playsound==1.2.2
plotly==5.17.0
preshed==3.0.9
prometheus-client==0.12.0
prompt-toolkit==3.0.22
Protego==0.2.1
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
PyDispatcher==2.0.6
Pygments==2.16.1
pyloco==0.0.139
Pympler==1.0.1
pynndescent==0.5.10
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySimpleGUI==4.56.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==22.3.0
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
readme-renderer==42.0
regex==2023.8.8
reportlab==3.6.3
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.7.1
seaborn==0.12.2
selenium==4.6.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==21.1.0
SimpleWebSocketServer==0.1.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.7
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
terminado==0.12.1
testpath==0.5.0
thinc==8.1.12
threadpoolctl==3.2.0
tldextract==3.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.1.1
transformers==4.33.2
trio==0.22.0
trio-websocket==0.9.2
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
umap-learn==0.5.4
urllib3==1.26.7
ushlex==0.99.1
validators==0.22.0
w3lib==2.0.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==4.0.4
wordcloud==1.9.2
wsproto==1.2.0
zipp==3.16.2
zope.interface==5.5.2
